Add data preprocessing
If you want to run this to test how it works I suggest not running it on the entire training set, e.g. do x1[0:100] instead of x1 and so on.

You will have to download test.json and train.json locally as they are too large to upload to Github.

There are a bunch of pre-processing options that we don't have to use but I am adding them in case we want to. You just need to set them to False to skip them.